api/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Apekun Dmitry <zasferg@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.2"
aiosqlite = "^0.20.0"
alembic = "^1.13.2"
uvicorn = "^0.30.6"
black = "^24.8.0"
pytest = "^8.3.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.34"}
pytest-asyncio = "^0.24.0"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.6.0"
alembic-autogen-check = "^1.1.1"
email-validator = "^2.2.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
